Moved tests to relevant file
